Refactor "Tips" tab to "About" and add app download links
Rename the "Tips" tab to "About" to better reflect its purpose.
This change includes updating the tab's title in the layout and
the visible text on the explore screen.

Additionally, the "explore.tsx" file now includes a footer
displaying "Made by Lugetech" with a link to their website.
The placeholder text for the note input on the "index.tsx"
screen has been updated to include download links for the
iOS and Android mobile applications, encouraging wider adoption.