Implement Welcome Screen and Image Compression
Implements a welcome screen that guides users through the app's functionality
and provides information on how to use it effectively. The screen includes
links to download the mobile app and offers a hint on entering a passphrase.

Adds image compression functionality to reduce the size of uploaded images
while maintaining their quality. This is achieved through the use of
expo-image-manipulator, which allows for progressive compression with
fallbacks to ensure that images are always compressed to a reasonable size.
The compression process is also accompanied by a progress indicator to
keep the user informed of the status.